Add edit-update function, and improve download function(for chrome, and solve confusion code problem)

+++ 2015.12.14/src/leibniz/hu/action/UserAction.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
 import java.util.List;
 
 import org.apache.commons.io.FileUtils;
@@ -105,11 +106,11 @@ public class UserAction extends ActionSupport implements ModelDriven<User>{
 	public String view(){
 		System.out.println("UserAction---view()");
 		UserDAO userDao = new UserDAO();
-		User user = userDao.getUserById(ServletActionContext.getRequest().getParameter("userID"));
+		User newuser = userDao.getUserById(ServletActionContext.getRequest().getParameter("userID"));
 		//将新的用户user放入值栈，以便回显
 		ValueStack vs = ServletActionContext.getContext().getValueStack();
 		vs.pop();
-		vs.push(user);
+		vs.push(newuser);
 		return "view";
 	}
 	
@@ -122,13 +123,18 @@ public class UserAction extends ActionSupport implements ModelDriven<User>{
 		//根据userID获取user属性（只是为了后缀名以及来判断是否有简历）
 		String userID = ServletActionContext.getRequest().getParameter("userID");
 		UserDAO userDao = new UserDAO();
-		User user = userDao.getUserById(userID);
+		User newuser = userDao.getUserById(userID);
 		//判断用户存在
-		if(null != user){
+		if(null != newuser){
 			//判断该用户有简历
-			if("1".equals(user.getResume())){
+			if("1".equals(newuser.getResume())){
 				//获取原文件名及后缀
-				String originFilename = user.getFilename();
+				try {
+					newuser.setFilename(new String(newuser.getFilename().getBytes("UTF-8"),"ISO8859-1"));
+				} catch (UnsupportedEncodingException e1) {
+					e1.printStackTrace();
+				}
+				String originFilename = newuser.getFilename();
 				String extension = originFilename.substring(originFilename.lastIndexOf("."));
 				//拼接成保存上传文件的文件名
 				String storedFilename = userID + extension;
@@ -136,11 +142,11 @@ public class UserAction extends ActionSupport implements ModelDriven<User>{
 				//创建文件输入流
 				try {
 					FileInputStream fis = new FileInputStream(new File(storedPath, storedFilename));
-					user.setDownloadStream(fis);
+					newuser.setDownloadStream(fis);
 					//加入到值栈中取代原有的user对象
 					ValueStack vs = ServletActionContext.getContext().getValueStack();
 					vs.pop();
-					vs.push(user);
+					vs.push(newuser);
 					return "down";
 				} catch (FileNotFoundException e) {
 					e.printStackTrace();
@@ -156,9 +162,75 @@ public class UserAction extends ActionSupport implements ModelDriven<User>{
 	 */
 	public String delete(){
 		System.out.println("UserAction---delete()");
-		UserDAO userDao = new UserDAO();
 		String userID = ServletActionContext.getRequest().getParameter("userID");
-		userDao.deleteByID(userID);
+		//查询到userID，并删除对应的附件
+		UserDAO userDao = new UserDAO();
+		User newuser = userDao.getUserById(userID);
+		//判断用户存在
+		if(null != newuser){
+			//判断该用户有简历
+			if("1".equals(newuser.getResume())){
+				//获取原文件名及后缀
+				String originFilename = newuser.getFilename();
+				String extension = originFilename.substring(originFilename.lastIndexOf("."));
+				//拼接成保存上传文件的文件名
+				String storedFilename = userID + extension;
+				String storedPath = ServletActionContext.getServletContext().getRealPath("upload");
+				new File(storedPath, storedFilename).delete();
+			}
+			userDao.deleteByID(userID);
+		}
+		return "relist";
+	}
+	
+	/**
+	 * 编辑用户，先查询数据库得到user，替代值栈，用于回显数据，并考虑处理兴趣爱好的字符串
+	 * @return
+	 */
+	public String edit(){
+		System.out.println("UserAction---edit()");
+		String userID = ServletActionContext.getRequest().getParameter("userID");
+		//查询到userID
+		UserDAO userDao = new UserDAO();
+		User newuser = userDao.getUserById(userID);
+		//查询得到的user取代值栈中的user，以便回显
+		ValueStack vs = ServletActionContext.getContext().getValueStack();
+		vs.pop();
+		vs.push(newuser);
+		//处理兴趣爱好，原为逗号隔开的字符串
+		String interest = newuser.getInterest();
+		if(null != interest && !("".equals(interest.trim()))){
+			//拆分成字符串数组并塞入request属性中
+			String[] interests = interest.split(", ");
+			ServletActionContext.getRequest().setAttribute("interests", interests);
+		}
+		return "edit";
+	}
+	
+	/**
+	 * 从编辑的jsp页面归来，将页面中的信息更新到数据库中
+	 * @return
+	 */
+	public String update(){
+		System.out.println("UserAction---update()");
+		UserDAO userDao = new UserDAO();
+		System.out.println(user);
+		if(null != user.getUploadContentType() && user.getUploadFileName().trim() != ""){
+			String originFileName = user.getUploadFileName();
+			String extension = originFileName.substring(originFileName.lastIndexOf("."));
+			String path = ServletActionContext.getServletContext().getRealPath("upload");
+			File file = new File(path, user.getUserID() + extension);
+			try {
+				//将上传的临时文件复制到指定的上传文件夹，并删除临时文件
+				FileUtils.copyFile(user.getUpload(), file);
+				user.getUpload().delete();
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+			user.setFilename(originFileName);
+			user.setResume("1");
+		}
+		userDao.updateUser(user);
 		return "relist";
 	}
 }
